feat: Doxygen comments for NavigationSystem
